// src/db/models/stripeConnect/transactions.ts
import {
  STRIPE_CONNECT_PAYMENT_STATUS,
  TRANSFER_STATUS,
} from "@/modules/stripe-connect/utils/stripe-connect.enum";
import mongoose from "mongoose";

const ObjectId = mongoose.Schema.Types.ObjectId;

export interface ITransaction {
  _id: mongoose.Types.ObjectId;
  stripeAccountId: string;
  stripeCustomerId: string;
  productRef?: mongoose.Types.ObjectId;
  amountPaid: number;
  stripePaymentIntentId: string;
  stripeChargeId?: string;
  companyRef: mongoose.Types.ObjectId;
  paymentStatus: STRIPE_CONNECT_PAYMENT_STATUS;
  orderPlacedAt?: Date;
  isRefundable: boolean;
  refunded: boolean;
  transferStatus: TRANSFER_STATUS;
}

export interface ITransactionDocument extends ITransaction, mongoose.Document {
  createdAt: Date;
  updatedAt: Date;
}

// ---------------------------------------------
// This schema is part of `stripe connect` flow
// ---------------------------------------------
export const TransactionSchema = new mongoose.Schema<ITransactionDocument>(
  {
    // Unique identifier for the connected Stripe account of the vendor.
    // This ID is generated by Stripe when a vendor successfully onboards to Stripe Connect.
    stripeAccountId: {
      type: String,
      required: true,
    },
    // Unique identifier for the customer on Stripe.
    // This ID represents the customer who made the purchase and can be used
    // for future transactions, refunds, or to access customer payment information.
    stripeCustomerId: {
      type: String,
      required: true,
    },
    productRef: {
      type: ObjectId,
      ref: "Products",
      required: false,
    },
    // This basically represents the final amount paid by the customer.
    amountPaid: {
      type: Number,
      required: true,
    },
    // Unique identifier for a specific payment intent on Stripe.
    // A PaymentIntent represents intent to collect payment from a customer.
    // It helps in tracking the lifecycle of the payment process.
    stripePaymentIntentId: {
      type: String,
      required: true,
    },
    // Unique identifier for a successful charge on Stripe.
    // A Charge represents a successful payment and is created when a PaymentIntent succeeds.
    // This field may be empty if the payment hasn't been processed yet.
    stripeChargeId: {
      type: String,
    },
    companyRef: {
      type: ObjectId,
      required: true,
      ref: "Company",
    },
    // Tracks the status of the payment made by customer.
    paymentStatus: {
      type: String,
      enum: Object.values(STRIPE_CONNECT_PAYMENT_STATUS),
      default: STRIPE_CONNECT_PAYMENT_STATUS.PENDING,
    },
    orderPlacedAt: {
      type: Date,
    },
    // Tracks whether the amount can be refunded to the customer
    isRefundable: {
      type: Boolean,
      default: true,
    },
    // Tracks whether the amount has been refunded to the customer
    refunded: {
      type: Boolean,
      default: false,
    },
    // The 'transferStatus' key is used to track the status of the money transfer to the vendor.
    // It indicates whether the transfer was successful or not after the refund period has ended.

    transferStatus: {
      type: String,
      enum: Object.values(TRANSFER_STATUS),
      default: TRANSFER_STATUS.PENDING,
    },
  },
  { timestamps: true },
);

export const StripeConnectTransactions = mongoose.model<ITransactionDocument>(
  "StripeConnectTransactions",
  TransactionSchema,
);